requirements.txt
beaker-pyteal==0.3.6
black==22.10.0
cffi==1.15.1
click==8.1.3
docstring-parser==0.14.1
Flask==2.2.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
py-algorand-sdk==1.19.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyteal==0.19.0
semantic-version==2.10.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
